requirements.txt
selenium
langchain
langchain-community
langchain-openai
unstructured
markdown
python-dotenv
# python.exe -m pip install --upgrade pip